finish consolidating parse paths
