+++ src/protocol/resp/src/request/get.rs
@@ -75,12 +75,12 @@ mod tests {
     fn parser() {
         let parser = RequestParser::new();
         assert_eq!(
-            parser.parse(b"get \"\0\r\n key\"\r\n").unwrap().into_inner(),
+            parser.parse(b"get 0\r\n").unwrap().into_inner(),
             Request::Get(GetRequest::new(b"0"))
         );
 
          assert_eq!(
-            parser.parse(b"get 0\r\n").unwrap().into_inner(),
+            parser.parse(b"get \"\0\r\n key\"\r\n").unwrap().into_inner(),
             Request::Get(GetRequest::new(b"\0\r\n key"))
         );
