+++ src/protocol/resp/src/request/set.rs
@@ -20,28 +20,6 @@ pub struct SetRequest {
     get_old: bool,
 }
 
-fn take_bulk_string(array: &mut Vec<Message>) -> Result<Box<[u8]>, ParseError> {
-    if let Message::BulkString(s) = array.remove(1) {
-        if s.inner.is_none() {
-            return Err(ParseError::Invalid);
-        }
-
-        let s = s.inner.unwrap();
-
-        Ok(s)
-    } else {
-        Err(ParseError::Invalid)
-    }
-}
-
-fn take_bulk_string_as_u64(array: &mut Vec<Message>) -> Result<u64, ParseError> {
-    let s = take_bulk_string(array)?;
-    std::str::from_utf8(&s)
-        .map_err(|_| ParseError::Invalid)?
-        .parse::<u64>()
-        .map_err(|_| ParseError::Invalid)
-}
-
 impl TryFrom<Message> for SetRequest {
     type Error = ParseError;
 
@@ -183,150 +161,6 @@ impl SetRequest {
     }
 }
 
-// this is to be called after parsing the command, so we do not match the verb
-pub fn parse(input: &[u8]) -> IResult<&[u8], SetRequest> {
-    let (input, _) = space1(input)?;
-    let (input, key) = string(input)?;
-    let (input, _) = space1(input)?;
-    let (input, value) = string(input)?;
-    let (mut input, _) = space0(input)?;
-
-    let mut expire_time = None;
-    let mut mode = SetMode::Set;
-    let mut get_old = false;
-
-    loop {
-        match string(input) {
-            Ok((i, b"EX")) => {
-                if expire_time.is_some() {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                let (i, _) = space1(i)?;
-
-                let (i, seconds) = parse_u64(i)?;
-
-                expire_time = Some(ExpireTime::Seconds(seconds));
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"PX")) => {
-                if expire_time.is_some() {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                let (i, _) = space1(i)?;
-
-                let (i, milliseconds) = parse_u64(i)?;
-
-                expire_time = Some(ExpireTime::Milliseconds(milliseconds));
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"EXAT")) => {
-                if expire_time.is_some() {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                let (i, _) = space1(i)?;
-
-                let (i, unix_seconds) = parse_u64(i)?;
-
-                expire_time = Some(ExpireTime::UnixSeconds(unix_seconds));
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"PXAT")) => {
-                if expire_time.is_some() {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                let (i, _) = space1(i)?;
-
-                let (i, unix_milliseconds) = parse_u64(i)?;
-
-                expire_time = Some(ExpireTime::UnixMilliseconds(unix_milliseconds));
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"KEEPTTL")) => {
-                if expire_time.is_some() {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                expire_time = Some(ExpireTime::KeepTtl);
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"NX")) => {
-                if mode != SetMode::Set {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                mode = SetMode::Add;
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"XX")) => {
-                if mode != SetMode::Set {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                mode = SetMode::Replace;
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((i, b"GET")) => {
-                if get_old {
-                    return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-                }
-
-                get_old = true;
-
-                let (i, _) = space0(i)?;
-
-                input = i;
-            }
-            Ok((_, b"")) => {
-                break;
-            }
-            Ok((i, _)) => {
-                return Err(nom::Err::Failure((i, nom::error::ErrorKind::Tag)));
-            }
-            Err(_) => {
-                break;
-            }
-        }
-    }
-
-    let (input, _) = crlf(input)?;
-
-    Ok((
-        input,
-        SetRequest {
-            key: key.to_owned().into_boxed_slice(),
-            value: value.to_owned().into_boxed_slice(),
-            expire_time,
-            mode,
-            get_old,
-        },
-    ))
-}
-
 impl Compose for SetRequest {
     fn compose(&self, session: &mut session::Session) {
         let mut alen = 3;
@@ -398,165 +232,6 @@ impl Compose for SetRequest {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    // #[test]
-    // fn parse_inline() {
-    //     // test parsing the command verb
-    //     assert_eq!(
-    //         inline_command(b"set key value\r\n"),
-    //         Ok((&b" key value\r\n"[..], Command::Set))
-    //     );
-
-    //     // test parsing the remainder of the request
-    //     assert_eq!(
-    //         set::parse(b" key value\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: None,
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             }
-    //         ))
-    //     );
-
-    //     // test parsing the entire request in one pass
-    //     assert_eq!(
-    //         inline_request(b"set key value\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: None,
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing with expire time in seconds
-    //     assert_eq!(
-    //         inline_request(b"set key value EX 300\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: Some(ExpireTime::Seconds(300)),
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing with expire time in milliseconds
-    //     assert_eq!(
-    //         inline_request(b"set key value PX 12345\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: Some(ExpireTime::Milliseconds(12345)),
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing with expire time in unix seconds
-    //     assert_eq!(
-    //         inline_request(b"set key value EXAT 1652370171\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: Some(ExpireTime::UnixSeconds(1652370171)),
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing with expire time in unix milliseconds
-    //     assert_eq!(
-    //         inline_request(b"set key value PXAT 1652370171000\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: Some(ExpireTime::UnixMilliseconds(1652370171000)),
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing with `KEEPTTL`
-    //     assert_eq!(
-    //         inline_request(b"set key value KEEPTTL\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: Some(ExpireTime::KeepTtl),
-    //                 mode: SetMode::Set,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing of a add
-    //     assert_eq!(
-    //         inline_request(b"set key value NX\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: None,
-    //                 mode: SetMode::Add,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing of a replace
-    //     assert_eq!(
-    //         inline_request(b"set key value XX\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: None,
-    //                 mode: SetMode::Replace,
-    //                 get_old: false,
-    //             })
-    //         ))
-    //     );
-
-    //     // test parsing of a set that returns the previous value
-    //     assert_eq!(
-    //         inline_request(b"set key value GET\r\n"),
-    //         Ok((
-    //             &b""[..],
-    //             Request::Set(SetRequest {
-    //                 key: b"key".to_vec().into_boxed_slice(),
-    //                 value: b"value".to_vec().into_boxed_slice(),
-    //                 expire_time: None,
-    //                 mode: SetMode::Set,
-    //                 get_old: true,
-    //             })
-    //         ))
-    //     );
-    // }
 
     #[test]
     fn parser() {
